requirements.txt
pytest
pybids
black
rich
sourcery
